Fix crash when loading bones due to falling objects
This was a "Restoring the wrong level." crash, caused because
there's an attempt to load the bones to a different level from the
player.

It is hard to make that case work, but fortuntely, the gameplay of
that case is kind-of awkward anyway, so it makes more sense to just
remove the possibility rather than trying to implement it.