Fix bug when kicking a throne from a square with items on it
In NetHack 3.4.3, the items can't fall because it doesn't know the
level they fall onto. That's specified using a dungeon and dungeon
level of 0. However, I accidentally changed the specification to
specify them as -1 instead (because they're stored in a coordinate
structure) and didn't change the test to match (it was still
looking for 0). This commit changes them to match properly.